# xapi-server/docs/swagger.yaml
definitions:
...
  controllers.ReqLogin:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  models.Result:
    properties:
      code:
        example: 0
        type: integer
      data:
        type: object
      message:
        example: 请求信息
        type: string
    type: object
  models.UserInfo:
    properties:
      email:
        type: string
      intro:
        type: string
      phone:
        type: string
      status:
        type: string
      uid:
        type: integer
      username:
        type: string
    type: object
info:
  contact: {}
paths:
...
  /login:
    post:
      consumes:
      - application/json
      description: 描述信息
      parameters:
      - description: login
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/controllers.ReqLogin'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Result'
      tags:
      - accounts
...
  /person/check:
    post:
      consumes:
      - application/json
      description: 资料检查
      parameters:
      - description: 用户验证
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.UserInfo'
      produces:
      - application/json
      summary: 个人资料检查,检查时只会一次传一个字段
      tags:
      - accounts
  /person/store:
    post:
      consumes:
      - application/json
      description: 描述信息
      parameters:
      - description: 用户验证
        in: body
        name: account
        required: true
        schema:
          $ref: '#/definitions/models.UserInfo'
      produces:
      - application/json
      summary: 个人资料修改
      tags:
      - accounts
...
swagger: "2.0"